old-tests/submission/W3C/domhtml/HTMLScriptElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>script</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script defer charset='utf-8' src="http://w3c-test.org/resources/idlharness.js" rel="help">
</script>
<link rel="author" title="W3C">
</head>
<body>
<script>
var domtest = async_test("The IDL attributes src, type, charset, and defer, each must reflect the respective content attributes of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcharset;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("script");
      assert_equals(nodeList.length,4);
      testNode = nodeList.item(2);
      vcharset = testNode.charset;

      assert_equals(vcharset,"utf-8");
       
    domtest.done();
}));

</script>
<p>The IDL attributes src, type, charset, and defer, each must reflect the respective content attributes of the same name.</p>
<div id="log"></div>
</body></html>
